requirements.txt
# Backend dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Claude Agent SDK
claude-agent-sdk>=0.0.3

# Codex SDK (for Codex agent)
codex-sdk-py>=0.0.3

# Development
ruff>=0.8.0
mypy>=1.13.0
